feat(participation): key the level by conversation, and let clients read it
The level is a rule about the room: one setting governs every agent in a
conversation, so it hangs off the conversation rather than off one
agent's instanceId. A GET is added because the control is shown to every
member and anyone may change it — a device-local guess goes stale the
moment someone else touches it, and today the app has no way to render
the level it is actually in.

The cooldown leaves the request shape. The hold becomes a harness
default derived from the conversation's own activity rather than a
number a client picks, and a body that still carries one is now rejected
so a stale build is visible instead of silently ignored.
