+++ src/api/v2/agents/agents.router.ts
@@ -3,11 +3,9 @@ import { requireAccount } from "@/middleware/auth";
 import {
   agentJoinLimiter,
   agentJoinStatusLimiter,
-  agentParticipationLimiter,
 } from "@/middleware/rateLimit";
 import { joinHandler } from "./handlers/join";
 import { joinStatusHandler } from "./handlers/join-status";
-import { participationHandler } from "./handlers/participation";
 
 export const agentsRouter = Router();
 
@@ -27,15 +25,4 @@ agentsRouter.get(
   "/join/:instanceId",
   agentJoinStatusLimiter,
   joinStatusHandler,
-);
-// Sets how much an agent may speak. `requireAccount` for the same reason as
-// /join: an account-less JWT is an authorization failure, not a stale token.
-// The product rule is that any conversation member may change the level, so
-// this deliberately has no owner gate; membership itself lives in the XMTP
-// group and is not verifiable here.
-agentsRouter.patch(
-  "/:instanceId/participation",
-  agentParticipationLimiter,
-  requireAccount,
-  participationHandler,
 );
